Move ActionCardHelpModal to a separate component in the components folder.
